#jira UETOOL-1513 Support for adaptive virtual texture. #rb none
#ROBOMERGE-OWNER: jeanmichel.dignard
#ROBOMERGE-AUTHOR: jeremy.moore
#ROBOMERGE-SOURCE: CL 14516203 in //UE4/Release-4.26/... via CL 14516229 via CL 14516235
#ROBOMERGE-BOT: TOOLS (Main -> Dev-Tools-Staging) (v749-14533626)

[CL 14549424 by jeremy moore in Dev-Tools-Staging branch]

+++ Engine/Source/Runtime/Renderer/Private/VT/RuntimeVirtualTextureProducer.h
@@ -30,7 +30,7 @@ public:
 	struct FTileEntry
 	{
 		FVTProduceTargetLayer Targets[RuntimeVirtualTexture::MaxTextureLayers];
-		uint32 vAddress = 0;
+		uint64 vAddress = 0;
 		uint8 vLevel = 0;
 	};
 
@@ -88,7 +88,7 @@ public:
 		const FVirtualTextureProducerHandle& ProducerHandle,
 		uint8 LayerMask,
 		uint8 vLevel,
-		uint32 vAddress,
+		uint64 vAddress,
 		EVTRequestPagePriority Priority
 	) override;
 
@@ -99,7 +99,7 @@ public:
 		const FVirtualTextureProducerHandle& ProducerHandle,
 		uint8 LayerMask,
 		uint8 vLevel,
-		uint32 vAddress,
+		uint64 vAddress,
 		uint64 RequestHandle,
 		const FVTProduceTargetLayer* TargetLayers
 	) override;
